RavenDB-19501 Changes after a review, moved ArchivedDataProcessingBehavior.cs and extracted BackgroundWorkParameters
