+++ src/Raven.Client/Documents/Operations/EssentialIndexInformation.cs
@@ -1,3 +1,4 @@
+using Raven.Client.Documents.DataArchival;
 using Raven.Client.Documents.Indexes;
 using Raven.Client.Documents.Operations.DataArchival;
