+++ src/Raven.Client/Documents/Subscriptions/SubscriptionCriteria.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Linq.Expressions;
 using Raven.Client.Documents.Conventions;
+using Raven.Client.Documents.DataArchival;
 using Raven.Client.Documents.Indexes;
 using Raven.Client.Documents.Operations.DataArchival;
 using Raven.Client.Documents.Session.Loaders;
