+++ src/Raven.Server/Documents/Indexes/IndexItem.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Raven.Client;
+using Raven.Client.Documents.DataArchival;
 using Raven.Client.Documents.Operations.DataArchival;
 using Raven.Server.Documents.TimeSeries;
 using Sparrow.Json;
