+++ src/Raven.Server/Utils/Monitoring/IndexMetrics.cs
@@ -6,6 +6,7 @@
 
 using System.Collections.Generic;
 using System.Linq;
+using Raven.Client.Documents.DataArchival;
 using Raven.Client.Documents.Indexes;
 using Raven.Client.Documents.Operations.DataArchival;
 using Sparrow.Json.Parsing;
