+++ test/SlowTests/Server/Documents/DataArchival/DataArchivalIndexingTests.cs
@@ -6,6 +6,7 @@ using FastTests;
 using FastTests.Utils;
 using Raven.Client;
 using Raven.Client.Documents;
+using Raven.Client.Documents.DataArchival;
 using Raven.Client.Documents.Indexes;
 using Raven.Client.Documents.Indexes.Counters;
 using Raven.Client.Documents.Indexes.TimeSeries;
@@ -188,7 +189,7 @@ public class DataArchivalIndexingTests : RavenTestBase
                     Count = g.Sum(x => x.Count)
                 };
 
-            ArchivedDataProcessingBehavior = Raven.Client.Documents.Operations.DataArchival.ArchivedDataProcessingBehavior.ArchivedOnly;
+            ArchivedDataProcessingBehavior = Raven.Client.Documents.DataArchival.ArchivedDataProcessingBehavior.ArchivedOnly;
         }
     }
     
@@ -406,7 +407,7 @@ public class DataArchivalIndexingTests : RavenTestBase
                 };
             
             Index(x => x.AddressText, FieldIndexing.Search);
-            ArchivedDataProcessingBehavior = Raven.Client.Documents.Operations.DataArchival.ArchivedDataProcessingBehavior.ArchivedOnly;
+            ArchivedDataProcessingBehavior = Raven.Client.Documents.DataArchival.ArchivedDataProcessingBehavior.ArchivedOnly;
         }
     }
 
@@ -497,7 +498,7 @@ public class DataArchivalIndexingTests : RavenTestBase
                 };
             
             Index(x => x.AddressText, FieldIndexing.Search);
-            ArchivedDataProcessingBehavior = Raven.Client.Documents.Operations.DataArchival.ArchivedDataProcessingBehavior.IncludeArchived;
+            ArchivedDataProcessingBehavior = Raven.Client.Documents.DataArchival.ArchivedDataProcessingBehavior.IncludeArchived;
         }
     }
      
@@ -697,7 +698,7 @@ public class DataArchivalIndexingTests : RavenTestBase
                             {
                                 CompanyName = LoadDocument<Company>(counter.Name).Name
                             });
-            ArchivedDataProcessingBehavior = Raven.Client.Documents.Operations.DataArchival.ArchivedDataProcessingBehavior.ExcludeArchived;
+            ArchivedDataProcessingBehavior = Raven.Client.Documents.DataArchival.ArchivedDataProcessingBehavior.ExcludeArchived;
         }
     }
     
@@ -733,7 +734,7 @@ public class DataArchivalIndexingTests : RavenTestBase
                                          Min = segment.Min,
                                          Sum = segment.Sum
                                      });
-            ArchivedDataProcessingBehavior = Raven.Client.Documents.Operations.DataArchival.ArchivedDataProcessingBehavior.ExcludeArchived;
+            ArchivedDataProcessingBehavior = Raven.Client.Documents.DataArchival.ArchivedDataProcessingBehavior.ExcludeArchived;
         }
     }
     
@@ -994,7 +995,7 @@ public class DataArchivalIndexingTests : RavenTestBase
             {
                 @"map('Companies', function (u){ return { Name: u.Name, Count: 1};})",
             };
-            ArchivedDataProcessingBehavior = Raven.Client.Documents.Operations.DataArchival.ArchivedDataProcessingBehavior.ArchivedOnly;
+            ArchivedDataProcessingBehavior = Raven.Client.Documents.DataArchival.ArchivedDataProcessingBehavior.ArchivedOnly;
         }
     }
     
@@ -1006,7 +1007,7 @@ public class DataArchivalIndexingTests : RavenTestBase
             {
                 @"map('Companies', function (u){ return { Name: u.Name, Count: 1};})",
             };
-            ArchivedDataProcessingBehavior = Raven.Client.Documents.Operations.DataArchival.ArchivedDataProcessingBehavior.IncludeArchived;
+            ArchivedDataProcessingBehavior = Raven.Client.Documents.DataArchival.ArchivedDataProcessingBehavior.IncludeArchived;
         }
     }
 
@@ -1170,7 +1171,7 @@ User: counter.DocumentId
 };
 })"
             };
-            ArchivedDataProcessingBehavior = Raven.Client.Documents.Operations.DataArchival.ArchivedDataProcessingBehavior.ExcludeArchived;
+            ArchivedDataProcessingBehavior = Raven.Client.Documents.DataArchival.ArchivedDataProcessingBehavior.ExcludeArchived;
         }
     }
